requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
attrs==21.2.0
click==8.0.1
fastapi==0.68.0
flake8==3.9.0
grpcio==1.39.0
grpcio-tools==1.39.0
h11==0.12.0
importlib-metadata==3.7.3
iniconfig==1.1.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pycodestyle==2.7.0
pydantic==1.8.2
pyflakes==2.3.0
pyparsing==2.4.7
pytest==6.2.4
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.15.0
zipp==3.4.1
